+++ refena/lib/src/provider/types/notifier_provider.dart
@@ -16,6 +16,7 @@ class NotifierProvider<N extends BaseSyncNotifier<T>, T>
     this._builder, {
     super.onChanged,
     super.debugLabel,
+    super.debugVisibleInGraph = true,
   });
 
   final N Function(Ref ref) _builder;