+++ refena/lib/src/provider/types/state_provider.dart
@@ -28,6 +28,7 @@ class StateProvider<T> extends BaseWatchableProvider<StateNotifier<T>, T>
     super.onChanged,
     String Function(T state)? describeState,
     String? debugLabel,
+    super.debugVisibleInGraph = true,
   })  : _describeState = describeState,
         super(debugLabel: debugLabel ?? 'StateProvider<$T>');